refactor(components): 🚚 organize into subdirectories
Reorganize the `components` directory to improve project structure and clarity by
grouping each component into its own subdirectory. Previously, all component `.tsx`
files were located directly in the root of the `components` folder. This change
introduces a more scalable and maintainable directory structure.

Changes made:
- Moved `Button.tsx` into a new `Button` directory.
- Moved `CountdownTimer.tsx` into a new `CountdownTimer` directory.
- Moved `Footer.tsx` into a new `Footer` directory.
- Moved `NumberBox.tsx` into a new `NumberBox` directory.
- Centralized component exports through a new `index.ts` file at the root of the
  `components` directory.

This refactor aims to streamline the development process by organizing components
into dedicated directories, facilitating easier navigation and potential future
expansion with additional files such as styles or tests specific to each component.
