+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './Button/Button';
+export * from './CountdownTimer/CountdownTimer';
+export * from './NumberBox/NumberBox';
+export * from './Footer/Footer';
